defguard/terraform/requirements.txt
boto3==1.34.0
pyjwt==2.9.0